refactor(backend): extract codex executable resolution
